refactor: rename tanstack query generator dir
