+++ packages/cli/bin.mjs
@@ -6,7 +6,7 @@
  * without a need `yarn install`
  */
 import { program } from './dist/bin.js';
-import plugin from './dist/generators/tanstack-react/plugin.js';
+import plugin from './dist/generators/tanstack-query-react/plugin.js';
 
 plugin.setupCommand(program);
