+++ packages/cli/src/generators/tanstack-query-react/generateCode.ts
