+++ packages/cli/src/generators/tanstack-query-react/plugin.ts
