+++ packages/cli/src/generators/tanstack-query-react/tanstack-query-react-client.spec.ts
