+++ packages/cli/src/generators/tanstack-query-react/ts-factory/getClientFactory.ts
