+++ packages/cli/src/generators/tanstack-query-react/ts-factory/getIndexFactory.ts
