+++ packages/cli/src/generators/tanstack-query-react/writeOpenAPIServices.ts
